services/voice-agent/requirements.txt
livekit-agents>=1.0.0
livekit-plugins-openai>=1.0.0
livekit-plugins-anthropic>=1.0.0
livekit-plugins-silero>=1.0.0
python-dotenv>=1.0.0
